src/chatbot/requirements.txt
# Streamlit and UI
streamlit>=1.28.0

# LangChain ecosystem for RAG
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6
langchain-pinecone>=0.1.0

# Vector database and embeddings
pinecone>=3.0.0
openai>=1.0.0

# LLM providers
anthropic>=0.8.0
google-generativeai>=0.3.0

# Environment and utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests
requests>=2.31.0
